move config files.
